Fix Grid Generation
The createGrid function was changed to use nested for loops to create
the grid by creating a column div and inserting each box div into the
column and repeating that for the total number of columns.

The css style of the various divs were changed to flexbox so that the
'picture' area is a set size, but the boxes change size as more/less
columns are entered.